pyproject.toml
[project]
name = "finalproject_magomedov_M25-555"
version = "0.1.0"
description = "Платформа для отслеживания и симуляции торговли валютами"
readme = "README.md"
authors = [
    {name = "Arkady Magomedov",email = "a.magomedov@live.com"}
]
requires-python = ">=3.10"
dependencies = [
    "prompt (>=0.4.1,<0.5.0)",
    "prettytable (>=3.16.0,<4.0.0)",
    "requests (>=2.32.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]

[tool.poetry]
packages = [
    { include = "valutatrade_hub" },
]
 
[tool.poetry.scripts]
project = "valutatrade_hub.cli.interface:main" 

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py312"
exclude = ["venv", ".venv", "dist", "__pycache__"]

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[dependency-groups]
dev = [
    "ruff (>=0.13.1,<0.14.0)"
]
